requirements.txt
django>=1.6,<1.7
cookiecutter>=0.7,<0.8
factory-boy>=2.3,<2.4
django-bootstrap-form>=3.1,<3.2